pyproject.toml
[project]
name = "remotepicam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "requests>=2.32.3",
]
